docs: Separate parser from harness
